requirements.txt
bluepy==1.3.0
btlewrap==0.0.10
miflora==0.7.1
RPi.GPIO==0.7.0
